move ProcessView and linux specific functionality to different file
