+++ elfcore/src/linux/mod.rs
@@ -1,4 +1,5 @@
 //! This module contains the functionality for
 //! gathering process information on Linux systems.
 
+pub mod memory;
 pub mod process;
